package stucture + quick fix for cross pngs paths
