+++ nomad_lab_visualizer/batch_update.py
@@ -1,4 +1,4 @@
-from include._geometry import make_hull, make_line
+from .geometry import make_hull, make_line
 
 
 def batch_update(self):
