+++ nomad_lab_visualizer/instantiate_widgets.py
@@ -274,7 +274,12 @@ def instantiate_widgets(self):
         ]
     )
 
-    file1 = open("./include/assets/cross.png", "rb")
+    import os
+    import nomad_lab_visualizer
+
+    file1 = open(
+        os.path.join(nomad_lab_visualizer.__path__[0], "assets/cross1.png"), "rb"
+    )
     image1 = file1.read()
     self.widg_img1 = widgets.Image(
         value=image1,
@@ -282,7 +287,9 @@ def instantiate_widgets(self):
         width=30,
         height=30,
     )
-    file2 = open("./include/assets/cross2.png", "rb")
+    file2 = open(
+        os.path.join(nomad_lab_visualizer.__path__[0], "assets/cross2.png"), "rb"
+    )
     image2 = file2.read()
     self.widg_img2 = widgets.Image(
         value=image2,
